+++ Controller/middleware/applyQueryString.js
@@ -1,5 +1,5 @@
 // __Module Definition__
-var mixin = module.exports = function () {
+var decorator = module.exports = function () {
   // Apply various options based on request query parameters.
   this.query(false, function (request, response, next) {
     var populate = request.query.populate;