+++ Controller/middleware/checkMethodSupported.js
@@ -1,5 +1,5 @@
 // __Module Definition__
-var mixin = module.exports = function () {
+var decorator = module.exports = function () {
   this.request(true, function (request, response, next) {
     var method = request.method === 'DELETE' ? 'del' : request.method.toLowerCase();
     if (request.baucis.controller.get(method) !== false) return next();