+++ Controller/middleware/initialize.js
@@ -1,5 +1,5 @@
 // __Module Definition__
-var mixin = module.exports = function () {
+var decorator = module.exports = function () {
   var controller = this;
 
   controller.request(true, function (request, response, next) {