+++ index.js
@@ -1,5 +1,7 @@
 // __Dependencies__
 var Api = require('./Api');
+var Release = require('./Release');
+var Controller = require('./Controller');
 
 // __Private Module Members__
 var instance = Api();
@@ -14,6 +16,11 @@ var baucis = module.exports = function (options) {
   previous.initialize();
   return previous;
 };
+
+// __Expose Modules__
+baucis.Api = Api;
+baucis.Release = Release;
+baucis.Controller = Controller;
 
 // __Public Methods__
 baucis.rest = function (options) {